Available seats refactored
