+++ app/concepts/tickets/representer.rb
@@ -10,9 +10,10 @@ module Tickets
       tickets.map do |ticket|
         {
             id: ticket.id,
-            date: ticket.price,
-            movie_id: ticket.type,
-            cinema_hall_id: ticket.reservation_id
+            price: ticket.price,
+            type: ticket.type,
+            seat: ticket.seat,
+            reservation_id: ticket.reservation_id
         }
       end
     end
