+++ app/controllers/tickets_controller.rb
@@ -10,14 +10,15 @@ class TicketsController < ApplicationController
   end
 
   def create
-    if Tickets::UseCases::SeatAvailable.new(ticket_params[:reservation_id], ticket_params[:seat]).call
+    if Tickets::UseCases::SeatAvailable.new(ticket_params[:reservation_id],ticket_params[:seat]).call
       ticket = Tickets::UseCases::Create.new.call(params: ticket_params)
       
       if ticket.valid?
+          update_screening_seats
           render json: ticket, status: :created
-        else
+      else
           render json: ticket.errors, status: :unprocessable_entity
-        end
+      end
     else
       render json: {status: "seat taken"}
     end
@@ -44,5 +45,13 @@ class TicketsController < ApplicationController
   def ticket_params
     params.require(:ticket).permit(:reservation_id, :type, :price, :seat)
   end
+
+  def update_screening_seats
+    reservation = Reservations::Repository.new.find_by(ticket_params[:reservation_id])
+    screening_id = reservation.screening_id
+    available_seats = Screenings::UseCases::FindAvailableSeats.new(screening_id).call
+    Screenings::UseCases::Update.new.call(id: screening_id, params: {available_seats: available_seats })
+  end
+
   
 end
